make calc_working_day module
